# Conflicts:
#	backend/src/main/scala/com/neo/sk/tank/core/UserActor.scala
#	shared/src/main/scala/com/neo/sk/tank/shared/protocol/TankGameEvent.scala

+++ frontend/src/main/scala/com/neo/sk/tank/front/pages/MainPage.scala
@@ -31,7 +31,7 @@ object MainPage extends PageSwitcher {
     case Nil => TankDemo.render
     case "playGame" :: playInfoSeq => PlayPage(playInfoSeq).render
     case "watchRecord":: rid :: wid :: f :: accessCode :: Nil => new ReplayPage(ReplayInfo(rid.toLong,wid.toLong, f.toInt, accessCode)).render
-    case "watchGame" :: roomId :: playerId :: accessCode ::Nil => new TankObservation(roomId.toLong, accessCode, Some(playerId.toLong)).render
+    case "watchGame" :: roomId :: playerId :: accessCode ::Nil => new TankObservation(roomId.toLong, accessCode, Some(playerId)).render
     case "watchGame" :: roomId :: accessCode :: Nil => new TankObservation(roomId.toLong, accessCode).render
     case "getGameRec" :: Nil => GameRecordPage.render
